Huge refactoring and cleanup.
In command mode, key presses can now be consumed directly and converted.
The decision to do so is now done at the highest possible level when all
modifications through layers, modifiers and second use have already been applied
to the keyboard_report.

Char strings or sequences of HID codes can be printed out as if they had been
entered on the keyboard. Both together allow a simplified macro handling
for both record and print as well as passhash output.

documentation