Keymap: One handed trackpoint operation but _NOT_ on thumb buttons.
This seemed most ergonomic, but has several problems:
Thumbs are DualUse, so while in mouse mode for example ESC to quit that mode is not available - can be overcome by any other key.
But using any modifier like Shift while in mouse-mode does not work, so move them all up to homerow and allow scrolling by pinky or thumb.